[MI-2589]: Fixed the bug of updating the redux state after login and refactor the code for redux of Jaas  (#11)
* [MI-2589]:Fixed the bug and refactor the code

* [MI-2589]:Refactoring the code

* [MI-2589]:Fixed self review fixes

* [MI-2589]:Fixed review fixes

* [MI-2589]:Fixed review comments
