--- webapp/src/jaas/action_types/index.ts
@@ -1,6 +1,0 @@
-
-import {id as pluginId} from '../../manifest';
-
-export default {
-    START_MEETING_WINDOW: pluginId + '_jaas/START_MEETING_WINDOW'
-};
